refactor(mobile): reuse connect and pairing components
