+++ apps/mobile/src/features/Ledger/LedgerPairing.container.tsx
@@ -1,77 +1,15 @@
-import React, { useRef } from 'react'
-import { View } from 'tamagui'
-import { router, useLocalSearchParams, useFocusEffect } from 'expo-router'
+import React from 'react'
 import type { DiscoveredDevice } from '@ledgerhq/device-management-kit'
-import { useLedgerConnection } from '@/src/features/Ledger/hooks/useLedgerConnection'
-import { PairingProgress } from '@/src/features/Ledger/components/PairingProgress'
-import { PairingError } from '@/src/features/Ledger/components/PairingError'
+import { LedgerPairing } from '@/src/features/Ledger/components/LedgerPairing'
 
 export const LedgerPairingContainer = () => {
-  const params = useLocalSearchParams<{
-    deviceData: string
-  }>()
-
-  const { connectToDevice, connectionError, clearError, isConnecting } = useLedgerConnection()
-  const hasPairingStarted = useRef(false)
-
-  const device: DiscoveredDevice | null = params.deviceData ? JSON.parse(params.deviceData) : null
-
-  const handleDeviceConnection = React.useCallback(async () => {
-    if (!device) {
-      return
-    }
-
-    const ledgerDevice = {
-      id: device.id,
-      name: device.name,
-      device,
-    }
-
-    const session = await connectToDevice(ledgerDevice)
-    if (session) {
-      router.replace({
-        pathname: '/import-signers/ledger-addresses',
-        params: {
-          deviceName: device.name,
-          sessionId: session,
-        },
-      })
-    }
-  }, [device, connectToDevice])
-
-  useFocusEffect(
-    React.useCallback(() => {
-      if (device && !hasPairingStarted.current) {
-        hasPairingStarted.current = true
-        handleDeviceConnection()
-      }
-    }, [device, handleDeviceConnection]),
-  )
-
-  const handleRetryPairing = async () => {
-    clearError()
-    await handleDeviceConnection()
+  const navigationConfig = {
+    pathname: '/import-signers/ledger-addresses',
+    getParams: (device: DiscoveredDevice, sessionId: string) => ({
+      deviceName: device.name,
+      sessionId,
+    }),
   }
 
-  if (connectionError && !isConnecting) {
-    return (
-      <View flex={1}>
-        <View flex={1} alignItems="center" justifyContent="center" paddingHorizontal="$4">
-          <PairingError
-            deviceName={device?.name || 'Unknown Device'}
-            errorMessage={connectionError.message}
-            onRetry={handleRetryPairing}
-          />
-        </View>
-      </View>
-    )
-  }
-
-  return (
-    <View flex={1}>
-      <View flex={1} alignItems="center" justifyContent="center" paddingHorizontal="$4">
-        <PairingProgress deviceName={device?.name || 'Unknown Device'} />
-      </View>
-    </View>
-  )
+  return <LedgerPairing navigationConfig={navigationConfig} />
 }
